backend/requirements.txt
boto3==1.34.0
PyJWT==2.8.0
botocore==1.34.0
requests==2.31.0
pytest==7.4.3
python-dateutil==2.8.2